utils: introduce heap_heap library
It tracks number of allocations and can help with detecting memory
leaks.

On Mac it was tested, on Linux was not, but most likely will work
the same.

The idea is that libc heap-related functions are expected to be
weak symbols. Meaning they can be redefined. The library simply
redefines the functions which students are expected to use and
tracks how many allocations remain unfreed.

hw2 shell: fix bigdata test

Mac 'wc -l' command somewhy prints some spaces before the number.
On Linux it doesn't happen, apparently. The patch trims the output
to make it look the same everywhere.

Thanks to Artem Bulgakov for the report.

hw2 shell: allow not to implement 'cd ~'

'~' is home directory of the current user. Fetching this location
is too cumbersome. Lets explicitly allow not to bother with it.

Thanks to Artem Chernitsa for the report.

hw2 shell: clarify 'exit' command

It needs to be supported in a special way. The patch also adds
new tests specifically for it.

Thanks to Robert Chen for the report.

g This is a combination of 7 commits.

[wip]

[wip]

[wip]

[wip]

15 points test passes

Restore checker

Remove test_cmd

Functional tests pass

Fix leaks

g This is a combination of 7 commits.

[wip]

[wip]

[wip]

[wip]

15 points test passes

Restore checker

Remove test_cmd

Functional tests pass

Fix leaks

[wip]

Handle status